+++ public/header.gohtml
@@ -1,0 +1,4 @@
+<meta charset="utf-8">
+<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+<meta name="author" content="Virtual Video Game Orchestra">
+<link rel="icon" href="images/favicons/favicon-32x32.png" sizes="32x32" type="image/png">
