+++ public/index.gohtml
@@ -1,0 +1,36 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+    {{ .Header }}
+    <title>Virtual Video Game Orchestra</title>
+    <meta name="description" content="">
+
+    <link rel="stylesheet" type="text/css" href="npm/bootstrap/dist/css/bootstrap.min.css">
+    <link rel="stylesheet" type="text/css" href="npm/@fortawesome/fontawesome-free/css/all.min.css">
+    <link rel="stylesheet" type="text/css" href="css/theme.css">
+
+    <script type="text/javascript" src="npm/jquery/dist/jquery.min.js"></script>
+    <script type="text/javascript" src="npm/bootstrap/dist/js/bootstrap.min.js"></script>
+    <script type="text/javascript" src="npm/@fortawesome/fontawesome-free/js/fontawesome.min.js"></script>
+</head>
+<body>
+{{ .NavBar }}
+<div class="container jumbotron bg-transparent">
+    <div class="bg-terminal">
+        <div class="p-3 thick-border rounded crt-grid">
+            <img src="images/logo.svg" class="logo-main btn bg-transparent"><a href="/" class=""></a>
+            <table class="m-auto">
+                <tr>
+                    <td class="btn btn-outline-light bg-transparent text-light btn-link text-monospace">
+                        <a href="https://www.facebook.com/groups/1080154885682377/" class="">
+                            Please join our facebook group!
+                        </a>
+                    </td>
+                </tr>
+            </table>
+        </div>
+    </div>
+</div>
+</body>
+</html>
