--- public/index.html
@@ -1,73 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta name="description" content="">
-    <meta name="author" content="Virtual Video Game Orchestra">
-
-    <title>Virtual Video Game Orchestra</title>
-
-    <link rel="icon" href="images/favicons/favicon-32x32.png" sizes="32x32" type="image/png">
-    <link rel="stylesheet" type="text/css" href="npm/bootstrap/dist/css/bootstrap.min.css">
-    <link rel="stylesheet" type="text/css" href="npm/@fortawesome/fontawesome-free/css/all.min.css">
-    <link rel="stylesheet" type="text/css" href="css/theme.css">
-
-    <script type="text/javascript" src="npm/jquery/dist/jquery.min.js"></script>
-    <script type="text/javascript" src="npm/bootstrap/dist/js/bootstrap.min.js"></script>
-    <script type="text/javascript" src="npm/@fortawesome/fontawesome-free/js/fontawesome.min.js"></script>
-</head>
-
-<div class="container">
-    <nav class="top-nav navbar navbar-expand-md navbar-dark bg-dark-blue-transparent fa-border m-2">
-        <a class="nav-link btn-outline-primary text-light" href="/">
-            <i class="fas fa-home fa-2x mr-2"></i>Virtual VGO
-        </a>
-        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarCollapse"
-                aria-controls="navbarCollapse" aria-expanded="false" aria-label="Toggle navigation">
-            <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse" id="navbarCollapse">
-            <ul class="navbar-nav container">
-                <li class="nav-item btn">
-                    <a class="nav-link btn-outline-primary text-monospace btn-link" href="/parts">
-                        Parts
-                    </a>
-                </li>
-            </ul>
-            <ul class="navbar-nav ml-auto">
-                <li class="nav-item">
-                    <a class="nav-link btn-outline-primary" href="https://www.facebook.com/groups/1080154885682377/">
-                        <i class="fab fa-facebook-square fa-2x"></i>
-                    </a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link btn-outline-primary" href="https://github.com/virtual-vgo/vvgo">
-                        <i class="fab fa-github-square fa-2x"></i>
-                    </a>
-                </li>
-            </ul>
-        </div>
-    </nav>
-</div>
-
-<body>
-<div class="container jumbotron bg-transparent">
-    <div class="bg-terminal">
-        <div class="p-3 thick-border rounded crt-grid">
-            <img src="images/logo.svg" class="logo-main btn bg-transparent"><a href="/" class=""></a>
-            <table class="m-auto">
-                <tr>
-                    <td class="btn btn-outline-light bg-transparent text-light btn-link text-monospace">
-                        <a href="https://www.facebook.com/groups/1080154885682377/" class="">
-                            Please join our facebook group!
-                        </a>
-                    </td>
-                </tr>
-            </table>
-        </div>
-    </div>
-</div>
-</body>
-</html>
