+++ public/navbar.gohtml
@@ -1,0 +1,32 @@
+<div class="container">
+    <nav class="top-nav navbar navbar-expand-md navbar-dark bg-dark-blue-transparent fa-border m-2">
+        <a class="nav-link btn-outline-primary text-light" href="/">
+            <i class="fas fa-home fa-2x mr-2"></i>Virtual VGO
+        </a>
+        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarCollapse"
+                aria-controls="navbarCollapse" aria-expanded="false" aria-label="Toggle navigation">
+            <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarCollapse">
+            <ul class="navbar-nav container">
+                <li class="nav-item btn">
+                    <a class="{{- if .PartsActive -}}nav-selected{{- end -}} nav-link btn-outline-primary text-monospace btn-link"
+                       href="/parts"> Parts
+                    </a>
+                </li>
+            </ul>
+            <ul class="navbar-nav ml-auto">
+                <li class="nav-item">
+                    <a class="nav-link btn-outline-primary" href="https://www.facebook.com/groups/1080154885682377/">
+                        <i class="fab fa-facebook-square fa-2x"></i>
+                    </a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link btn-outline-primary" href="https://github.com/virtual-vgo/vvgo">
+                        <i class="fab fa-github-square fa-2x"></i>
+                    </a>
+                </li>
+            </ul>
+        </div>
+    </nav>
+</div>
